:sparkles: integrate jotai for global sidebar menu state management
