+++ package.json
@@ -17,6 +17,7 @@
   "dependencies": {
     "i18next": "^23.12.2",
     "i18next-resources-to-backend": "^1.2.1",
+    "jotai": "^2.9.3",
     "next": "14.2.5",
     "next-i18n-router": "^5.5.1",
     "react": "^18",
