+++ src/store/index.ts
@@ -1,0 +1,3 @@
+import { atom } from 'jotai';
+
+export const isSidebarMenuOpen = atom(false);
